pyproject.toml
[tool.poetry]
name = "prdagentsystem"
version = "0.1.0"
description = ""
authors = ["NIKHIL KUMAR REDDY <nikhilkumar.m@skil.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
streamlit = "^1.40.2"
crewai = "^0.83.0"
crewai-tools = "^0.14.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
